Added new icons for museic box and done museic box specific changes (except the record button)

+++ ListeningAloudiPadHarp/src/Sampler.h
@@ -75,7 +75,7 @@ public:
 	
 	// give this function a normalized
 	// value, and it gives back a note number
-	int valueToNote(float y);
+	int valueToNote(int y);
 	
 	// converts a note number to sample playback speed
 	float noteToSpeed(int note);
